+++ src/lagrangeinterpolator.cpp
@@ -10,7 +10,7 @@ HdrImage LaGrangeInterpolator::takeStep() {
 	return interpolation;
 }
 
-HdrImage LaGrangeInterpolator::calculateInterpolationOn(int observation) {
+HdrImage LaGrangeInterpolator::calculateInterpolationOn(float observation) {
 	int numberOfObservations = observations_.size();
 	int middle = std::floor(numberOfObservations/2);
 	double medianObservation = observations_[middle];
