+++ src/lagrangeinterpolator.h
@@ -8,7 +8,7 @@
 
 class LaGrangeInterpolator : public Interpolator {
 public:
-	HdrImage calculateInterpolationOn(int observation);
+	HdrImage calculateInterpolationOn(float observation);
 	HdrImage takeStep();
 };
